GHI-352 Extract class
