+++ SafeUIKit/SafeUIKit/Components/TransactionFeeView/TransactionFeeCell.swift
@@ -4,28 +4,9 @@
 
 import UIKit
 
-public class TransactionFeeCell: UITableViewCell {
+public class TransactionFeeCell: ContainerCell {
 
     public let transactionFeeView = TransactionFeeView()
-
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        commonInit()
-    }
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        commonInit()
-    }
-
-    private func commonInit() {
-        transactionFeeView.translatesAutoresizingMaskIntoConstraints = false
-        contentView.addSubview(transactionFeeView)
-        NSLayoutConstraint.activate([
-            transactionFeeView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            transactionFeeView.topAnchor.constraint(equalTo: contentView.topAnchor),
-            transactionFeeView.rightAnchor.constraint(equalTo: contentView.rightAnchor),
-            transactionFeeView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)])
-    }
+    public override var cellContentView: UIView { return transactionFeeView }
 
 }
